refactor: move lock file into pact broker module
